# apps/backend/openapi.yaml
openapi: 3.0.3
info:
  title: TaskMind Backend API
  version: 0.1.0
  description: |
    Initial v1 contract draft for TaskMind backend task and project management endpoints.
servers:
  - url: http://localhost:8080
paths:
  /v1/tasks:
    post:
      summary: Create a task
      operationId: createTask
      tags: [Tasks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '201':
          description: Task created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/ValidationError'
        '403':
          description: User is not a member of the project
    get:
      summary: List tasks
      operationId: listTasks
      tags: [Tasks]
      parameters:
        - in: query
          name: userId
          schema:
            type: string
            format: uuid
        - in: query
          name: status
          schema:
            $ref: '#/components/schemas/TaskStatus'
        - in: query
          name: overdueOnly
          schema:
            type: boolean
            default: false
        - in: query
          name: page
          schema:
            type: integer
            minimum: 0
            default: 0
        - in: query
          name: size
          schema:
            type: integer
            minimum: 1
            default: 20
      responses:
        '200':
          description: Task list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/ValidationError'

  /v1/tasks/{id}:
    patch:
      summary: Update mutable task fields
      operationId: updateTask
      tags: [Tasks]
      parameters:
        - $ref: '#/components/parameters/TaskIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskRequest'
      responses:
        '200':
          description: Task updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/ValidationError'
        '403':
          description: User is not a member of the project
        '404':
          description: Task not found

  /v1/tasks/{id}/status:
    patch:
      summary: Update only task status
      operationId: updateTaskStatus
      tags: [Tasks]
      parameters:
        - $ref: '#/components/parameters/TaskIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskStatusRequest'
      responses:
        '200':
          description: Task status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          description: Task not found

  /v1/tasks/{id}/completion:
    get:
      summary: Get completion state for a task
      operationId: getTaskCompletion
      tags: [Tasks]
      parameters:
        - $ref: '#/components/parameters/TaskIdPath'
      responses:
        '200':
          description: Completion details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskCompletionResponse'
        '404':
          description: Task not found

  /v1/tasks/{id}/archive:
    patch:
      summary: Archive task (soft lifecycle transition)
      operationId: archiveTask
      tags: [Tasks]
      parameters:
        - $ref: '#/components/parameters/TaskIdPath'
      responses:
        '200':
          description: Task archived
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          description: Task not found

  /v1/projects:
    post:
      summary: Create a project
      operationId: createProject
      tags: [Projects]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        '201':
          description: Project created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          description: Project key already exists
    get:
      summary: List projects
      operationId: listProjects
      tags: [Projects]
      parameters:
        - in: query
          name: includeArchived
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Project list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Project'

  /v1/projects/{id}:
    get:
      summary: Get project by id
      operationId: getProject
      tags: [Projects]
      parameters:
        - $ref: '#/components/parameters/ProjectIdPath'
      responses:
        '200':
          description: Project details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '404':
          description: Project not found
    patch:
      summary: Update mutable project fields
      operationId: updateProject
      tags: [Projects]
      parameters:
        - $ref: '#/components/parameters/ProjectIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProjectRequest'
      responses:
        '200':
          description: Project updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          description: Project not found
        '409':
          description: Project key already exists


  /v1/projects/{projectId}/members:
    post:
      summary: Add project member
      operationId: addProjectMember
      tags: [Projects]
      parameters:
        - $ref: '#/components/parameters/ProjectIdPathAsProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddProjectMemberRequest'
      responses:
        '201':
          description: Member added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectMembership'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          description: Membership already exists
    get:
      summary: List project members
      operationId: listProjectMembers
      tags: [Projects]
      parameters:
        - $ref: '#/components/parameters/ProjectIdPathAsProjectId'
      responses:
        '200':
          description: Project memberships
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProjectMembership'

  /v1/projects/{projectId}/members/{userId}:
    delete:
      summary: Remove project member
      operationId: removeProjectMember
      tags: [Projects]
      parameters:
        - $ref: '#/components/parameters/ProjectIdPathAsProjectId'
        - $ref: '#/components/parameters/UserIdPath'
      responses:
        '204':
          description: Member removed

  /v1/projects/{id}/archive:
    patch:
      summary: Archive project
      operationId: archiveProject
      tags: [Projects]
      parameters:
        - $ref: '#/components/parameters/ProjectIdPath'
      responses:
        '200':
          description: Project archived
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '404':
          description: Project not found

  /v1/tasks/{id}/children:
    get:
      summary: List direct child tasks
      tags: [Tasks]
      responses:
        '200': {description: Child tasks}
  /v1/tasks/{id}/ancestors:
    get:
      summary: List task ancestors
      tags: [Tasks]
      responses:
        '200': {description: Ancestor tasks}
  /v1/tasks/{id}/links:
    get:
      summary: List cross-task links
      tags: [Tasks]
      responses:
        '200': {description: Task links}
    post:
      summary: Create a cross-task link
      tags: [Tasks]
      responses:
        '201': {description: Task link created}
  /v1/task-links/{linkId}:
    delete:
      summary: Delete a cross-task link
      tags: [Tasks]
      responses:
        '204': {description: Task link deleted}
  /v1/projects/{id}/releases:
    get:
      summary: Summarize project releases
      tags: [Projects]
      responses:
        '200': {description: Release summary}

components:
  parameters:
    TaskIdPath:
      in: path
      name: id
      required: true
      schema:
        type: string
        format: uuid
    ProjectIdPath:
      in: path
      name: id
      required: true
      schema:
        type: string
        format: uuid
    ProjectIdPathAsProjectId:
      in: path
      name: projectId
      required: true
      schema:
        type: string
        format: uuid
    UserIdPath:
      in: path
      name: userId
      required: true
      schema:
        type: string
        format: uuid

  responses:
    ValidationError:
      description: Input validation failure
      content:
        application/json:
          schema:
            type: object
            additionalProperties: true

  schemas:
    Project:
      type: object
      required:
        - id
        - name
        - key
        - ownerUserId
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 200
        key:
          type: string
          maxLength: 20
        description:
          type: string
          nullable: true
        ownerUserId:
          type: string
          format: uuid
        archivedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    CreateProjectRequest:
      type: object
      required:
        - name
        - key
        - ownerUserId
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        key:
          type: string
          minLength: 1
          maxLength: 20
        description:
          type: string
          nullable: true
        ownerUserId:
          type: string
          format: uuid

    UpdateProjectRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 200
        key:
          type: string
          maxLength: 20
        description:
          type: string
          nullable: true


    AddProjectMemberRequest:
      type: object
      required:
        - userId
        - role
      properties:
        userId:
          type: string
          format: uuid
        role:
          $ref: '#/components/schemas/ProjectMembershipRole'

    ProjectMembership:
      type: object
      required:
        - projectId
        - userId
        - role
      properties:
        projectId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        role:
          $ref: '#/components/schemas/ProjectMembershipRole'

    ProjectMembershipRole:
      type: string
      enum:
        - OWNER
        - ADMIN
        - MEMBER
        - VIEWER

    Task:
      type: object
      required:
        - id
        - userId
        - title
        - status
        - priority
        - source
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
          nullable: true
        taskKey:
          type: string
          nullable: true
        assigneeId:
          type: string
          format: uuid
          nullable: true
        parentTaskId:
          type: string
          format: uuid
          nullable: true
        taskLevel:
          $ref: '#/components/schemas/TaskLevel'
        taskType:
          $ref: '#/components/schemas/TaskType'
        storyPoints:
          type: integer
          minimum: 0
          nullable: true
        releaseVersion:
          type: string
          nullable: true
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/TaskStatus'
        priority:
          type: integer
          minimum: 1
          maximum: 4
        dueAt:
          type: string
          format: date-time
          nullable: true
        durationMinutes:
          type: integer
          minimum: 1
          nullable: true
        energyLevel:
          $ref: '#/components/schemas/EnergyLevel'
        source:
          $ref: '#/components/schemas/TaskSource'
        confidence:
          type: number
          format: double
          minimum: 0
          maximum: 1
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    CreateTaskRequest:
      type: object
      required:
        - userId
        - title
        - status
        - priority
        - source
      properties:
        userId:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
          nullable: true
        assigneeId:
          type: string
          format: uuid
          nullable: true
        parentTaskId:
          type: string
          format: uuid
          nullable: true
        taskLevel:
          $ref: '#/components/schemas/TaskLevel'
        taskType:
          $ref: '#/components/schemas/TaskType'
        storyPoints:
          type: integer
          minimum: 0
          nullable: true
        releaseVersion:
          type: string
          nullable: true
        title:
          type: string
          minLength: 1
        description:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/TaskStatus'
        priority:
          type: integer
          minimum: 1
          maximum: 4
        dueAt:
          type: string
          format: date-time
          nullable: true
        durationMinutes:
          type: integer
          minimum: 1
          nullable: true
        energyLevel:
          $ref: '#/components/schemas/EnergyLevel'
        source:
          $ref: '#/components/schemas/TaskSource'
        confidence:
          type: number
          format: double
          minimum: 0
          maximum: 1
          nullable: true

    UpdateTaskRequest:
      type: object
      properties:
        version:
          type: integer
          format: int64
          nullable: true
        projectId:
          type: string
          format: uuid
          nullable: true
        assigneeId:
          type: string
          format: uuid
          nullable: true
        parentTaskId:
          type: string
          format: uuid
          nullable: true
        taskLevel:
          $ref: '#/components/schemas/TaskLevel'
        taskType:
          $ref: '#/components/schemas/TaskType'
        storyPoints:
          type: integer
          minimum: 0
          nullable: true
        releaseVersion:
          type: string
          nullable: true
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/TaskStatus'
        priority:
          type: integer
          minimum: 1
          maximum: 4
        dueAt:
          type: string
          format: date-time
          nullable: true
        durationMinutes:
          type: integer
          minimum: 1
          nullable: true
        energyLevel:
          $ref: '#/components/schemas/EnergyLevel'

    UpdateTaskStatusRequest:
      type: object
      required:
        - status
      properties:
        status:
          $ref: '#/components/schemas/TaskStatus'

    TaskCompletionResponse:
      type: object
      required:
        - id
        - status
        - completed
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/TaskStatus'
        completed:
          type: boolean
        updatedAt:
          type: string
          format: date-time

    TaskStatus:
      type: string
      enum:
        - TODO
        - IN_PROGRESS
        - DONE
        - ARCHIVED

    TaskSource:
      type: string
      enum:
        - MANUAL
        - AI_CAPTURE
        - AI_PLAN
        - AI_RESCHEDULE

    EnergyLevel:
      type: string
      enum:
        - LOW
        - MEDIUM
        - HIGH

    TaskLevel:
      type: string
      enum: [EPIC, STORY, TASK, SUBTASK]
    TaskType:
      type: string
      enum: [EPIC, STORY, TASK, BUG, SUBTASK, MILESTONE]
    TaskLinkType:
      type: string
      enum: [BLOCKS, RELATES_TO, DUPLICATES]